+++ workspaces/redhat-argocd/plugins/argocd/src/components/DeploymentLifeCycle/sidebar/resources/resource/RolloutMetadata.tsx
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 import React from 'react';
-import { Resource } from '../../../../../types/application';
+import { Resource } from '@backstage-community/plugin-redhat-argocd-common';
 import { useArgoResources } from '../../rollouts/RolloutContext';
 import Rollout from '../../rollouts/Rollout';
 import { RolloutUI } from '../../../../../types/revision';